requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
numpy==2.1.3
opencv-python==4.10.0.84
sqlparse==0.5.2
